Make branch name from /ping use one at startup
Before, since getCurrentBranchName() was computed at command execution
time, the branch name returned would be whatever the repository happened
to be checked out to, which defeats the purpose of the "Branch" field,
which should describe the version of the bot currently running. Since
the program is already in main memory when running, checking out the
repository shouldn't affect the bot's runtime, so it should be
associated with the branch at startup time and not real time.
